pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-triage-env"
version = "0.1.0"
description = "An OpenEnv environment for customer support ticket triage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openai>=1.40.0",
  "openenv-core>=0.2.0",
  "pydantic>=2.8.0",
  "uvicorn>=0.30.0",
]

[project.scripts]
server = "support_triage_env.server.app:main"

[tool.setuptools.packages.find]
include = ["support_triage_env*"]
